# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vyapaar AI API Specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [crm]
      summary: List Customers
      description: Retrieves customers associated with the active organization.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [crm]
      summary: Create Customer
      description: Adds a new customer profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /leads:
    get:
      operationId: listLeads
      tags: [crm]
      summary: List Leads
      description: Retrieves sales leads and pipelines.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [crm]
      summary: Create Lead
      description: Inserts a new lead card.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200":
          description: Lead created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{leadId}/status:
    put:
      operationId: updateLeadStatus
      tags: [crm]
      summary: Update Lead Pipeline Stage
      description: Updates status stage of a sales lead.
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required: [status]
      responses:
        "200":
          description: Lead status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /inquiries:
    get:
      operationId: listInquiries
      tags: [crm]
      summary: List Public Site Inquiries
      description: Retrieves contact submissions.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inquiries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: submitInquiry
      tags: [crm]
      summary: Public Submit Inquiry
      description: Captures inquiries submitted by external clients on the business mini website.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orgId:
                  type: integer
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                message:
                  type: string
              required: [orgId, name]
      responses:
        "200":
          description: Inquiry registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /tasks:
    get:
      operationId: listTasks
      tags: [crm]
      summary: List Tasks
      description: Lists follow-up operational tasks.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [crm]
      summary: Create Task
      description: Creates a follow-up task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        branchId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
      required: [id, orgId, name]
    CustomerInput:
      type: object
      properties:
        orgId:
          type: integer
        branchId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
      required: [orgId, name]
    Lead:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        branchId:
          type: integer
        customerId:
          type: integer
        source:
          type: string
        status:
          type: string
        notes:
          type: string
        assignedTo:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, orgId, status]
    LeadInput:
      type: object
      properties:
        orgId:
          type: integer
        branchId:
          type: integer
        customerId:
          type: integer
        source:
          type: string
        status:
          type: string
        notes:
          type: string
        assignedTo:
          type: string
      required: [orgId]
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        branchId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, orgId, name]
    Task:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        branchId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date-time
        assignedTo:
          type: string
      required: [id, orgId, title, status]
    TaskInput:
      type: object
      properties:
        orgId:
          type: integer
        branchId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date-time
        assignedTo:
          type: string
      required: [orgId, title]
...